requirements.txt
# Database
psycopg2-binary==2.9.9

# Environment variables
python-dotenv==1.0.0

# Telegram
telethon==1.34.0

# Async support
asyncio==3.4.3

